backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain
langchain-community
langchain-openai
openai>=0.27.0
faiss-cpu
tiktoken
cohere

